requirements.txt
transformers
datasets
pillow
numpy
